+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2.3.0
 
 * Fix: shouldn't auto play when seeking in pause state
+* Fix: show the first frame when video loaded and Play() not called
 * keep screen on (disable screensaver) while playing video.
 
 ## 2.2.0